requirements.txt
flask
opencv-python
numpy
Pillow
ultralytics
mediapipe
gunicorn
